Addded reviews ratings, subscrption ...

+++ src/database/migrations/20220913072337-create-review.js
@@ -21,7 +21,7 @@ module.exports = {
         allowNull: false,
       },
       rate: {
-        type: Sequelize.DECIMAL(2, 1),
+        type: Sequelize.INTEGER,
         allowNull: false,
       },
       createdAt: {
